+++ commercetools/commercetools-sdk-java-api/src/main/java-generated/com/commercetools/api/models/message/MessageBuilder.java
@@ -287,6 +287,10 @@ public class MessageBuilder {
         return com.commercetools.api.models.message.OrderPaymentStateChangedMessageBuilder.of();
     }
 
+    public com.commercetools.api.models.message.OrderPurchaseOrderNumberSetMessageBuilder orderPurchaseOrderNumberSetBuilder() {
+        return com.commercetools.api.models.message.OrderPurchaseOrderNumberSetMessageBuilder.of();
+    }
+
     public com.commercetools.api.models.message.OrderReturnShipmentStateChangedMessageBuilder orderReturnShipmentStateChangedBuilder() {
         return com.commercetools.api.models.message.OrderReturnShipmentStateChangedMessageBuilder.of();
     }