+++ commercetools/commercetools-sdk-java-api/src/main/java-generated/com/commercetools/api/models/order_edit/StagedOrderBuilder.java
@@ -125,6 +125,9 @@ public class StagedOrderBuilder implements Builder<StagedOrder> {
     private java.util.List<com.commercetools.api.models.order.ReturnInfo> returnInfo;
 
     @Nullable
+    private String purchaseOrderNumber;
+
+    @Nullable
     private java.util.List<com.commercetools.api.models.cart.DiscountCodeInfo> discountCodes;
 
     @Deprecated
@@ -831,6 +834,15 @@ public class StagedOrderBuilder implements Builder<StagedOrder> {
     }
 
     /**
+     *  <p>The Purchase Order Number is typically set by the Buyer on a QuoteRequest to track the purchase order during the quote and order flow.</p>
+     */
+
+    public StagedOrderBuilder purchaseOrderNumber(@Nullable final String purchaseOrderNumber) {
+        this.purchaseOrderNumber = purchaseOrderNumber;
+        return this;
+    }
+
+    /**
      *
      */
 
@@ -1301,6 +1313,11 @@ public class StagedOrderBuilder implements Builder<StagedOrder> {
     }
 
     @Nullable
+    public String getPurchaseOrderNumber() {
+        return this.purchaseOrderNumber;
+    }
+
+    @Nullable
     public java.util.List<com.commercetools.api.models.cart.DiscountCodeInfo> getDiscountCodes() {
         return this.discountCodes;
     }
@@ -1382,9 +1399,9 @@ public class StagedOrderBuilder implements Builder<StagedOrder> {
             orderNumber, customerId, customerEmail, anonymousId, businessUnit, store, lineItems, customLineItems,
             totalPrice, taxedPrice, taxedShippingPrice, shippingAddress, billingAddress, shippingMode, shipping,
             taxMode, taxRoundingMode, customerGroup, country, orderState, state, shipmentState, paymentState,
-            shippingInfo, syncInfo, returnInfo, discountCodes, lastMessageSequenceNumber, cart, quote, custom,
-            paymentInfo, locale, inventoryMode, origin, taxCalculationMode, shippingRateInput, itemShippingAddresses,
-            refusedGifts);
+            shippingInfo, syncInfo, returnInfo, purchaseOrderNumber, discountCodes, lastMessageSequenceNumber, cart,
+            quote, custom, paymentInfo, locale, inventoryMode, origin, taxCalculationMode, shippingRateInput,
+            itemShippingAddresses, refusedGifts);
     }
 
     /**
@@ -1395,9 +1412,9 @@ public class StagedOrderBuilder implements Builder<StagedOrder> {
             orderNumber, customerId, customerEmail, anonymousId, businessUnit, store, lineItems, customLineItems,
             totalPrice, taxedPrice, taxedShippingPrice, shippingAddress, billingAddress, shippingMode, shipping,
             taxMode, taxRoundingMode, customerGroup, country, orderState, state, shipmentState, paymentState,
-            shippingInfo, syncInfo, returnInfo, discountCodes, lastMessageSequenceNumber, cart, quote, custom,
-            paymentInfo, locale, inventoryMode, origin, taxCalculationMode, shippingRateInput, itemShippingAddresses,
-            refusedGifts);
+            shippingInfo, syncInfo, returnInfo, purchaseOrderNumber, discountCodes, lastMessageSequenceNumber, cart,
+            quote, custom, paymentInfo, locale, inventoryMode, origin, taxCalculationMode, shippingRateInput,
+            itemShippingAddresses, refusedGifts);
     }
 
     public static StagedOrderBuilder of() {
@@ -1439,6 +1456,7 @@ public class StagedOrderBuilder implements Builder<StagedOrder> {
         builder.shippingInfo = template.getShippingInfo();
         builder.syncInfo = template.getSyncInfo();
         builder.returnInfo = template.getReturnInfo();
+        builder.purchaseOrderNumber = template.getPurchaseOrderNumber();
         builder.discountCodes = template.getDiscountCodes();
         builder.lastMessageSequenceNumber = template.getLastMessageSequenceNumber();
         builder.cart = template.getCart();