+++ commercetools/commercetools-sdk-java-api/src/main/java-generated/com/commercetools/api/models/quote/Quote.java
@@ -313,6 +313,13 @@ public interface Quote extends BaseResource, com.commercetools.api.models.Domain
     public StateReference getState();
 
     /**
+     *  <p>The Purchase Order Number is typically set by the Buyer on a QuoteRequest to track the purchase order during the quote and order flow.</p>
+     */
+
+    @JsonProperty("purchaseOrderNumber")
+    public String getPurchaseOrderNumber();
+
+    /**
      *  <p>The BusinessUnit for the Quote.</p>
      */
     @Valid
@@ -399,6 +406,8 @@ public interface Quote extends BaseResource, com.commercetools.api.models.Domain
 
     public void setState(final StateReference state);
 
+    public void setPurchaseOrderNumber(final String purchaseOrderNumber);
+
     public void setBusinessUnit(final BusinessUnitKeyReference businessUnit);
 
     public static Quote of() {
@@ -441,6 +450,7 @@ public interface Quote extends BaseResource, com.commercetools.api.models.Domain
         instance.setCustom(template.getCustom());
         instance.setQuoteState(template.getQuoteState());
         instance.setState(template.getState());
+        instance.setPurchaseOrderNumber(template.getPurchaseOrderNumber());
         instance.setBusinessUnit(template.getBusinessUnit());
         return instance;
     }