+++ changes.md
@@ -6,3 +6,26 @@
 - type `IronMqDestination` is removed
 </details>
 
+
+<details>
+<summary>Added Type(s)</summary>
+
+- added type `OrderPurchaseOrderNumberSetMessage`
+- added type `OrderPurchaseOrderNumberSetMessagePayload`
+- added type `StagedOrderSetPurchaseOrderNumberAction`
+- added type `OrderSetPurchaseOrderNumberAction`
+</details>
+
+
+<details>
+<summary>Added Property(s)</summary>
+
+- added property `purchaseOrderNumber` to type `StagedOrder`
+- added property `purchaseOrderNumber` to type `Order`
+- added property `purchaseOrderNumber` to type `OrderFromCartDraft`
+- added property `purchaseOrderNumber` to type `QuoteRequest`
+- added property `purchaseOrderNumber` to type `QuoteRequestDraft`
+- added property `purchaseOrderNumber` to type `Quote`
+- added property `purchaseOrderNumber` to type `StagedQuote`
+</details>
+